Crate linfa_linear

linfa-linear aims to provide pure Rust implementations of popular linear regression algorithms.

The Big Picture

linfa-linear is a crate in the linfa ecosystem, an effort to create a toolkit for classical Machine Learning implemented in pure Rust, akin to Python's scikit-learn.

Current state

linfa-linear currently provides an implementation of the following regression algorithms:

Ordinary Least Squares

Generalized Linear Models (GLM)

Examples

There is an usage example in the examples/ directory. To run, use:

$ cargo run --features openblas --example diabetes
$ cargo run --example glm

Structs

FittedLinearRegression
A fitted linear regression model which can be used for making predictions.

FittedTweedieRegressor
Fitted Tweedie regressor model for scoring

LinearRegression
An ordinary least squares linear regression model.

TweedieRegressor
Generalized Linear Model (GLM) with a Tweedie distribution

Enums

LinearError
An error when modeling a Linear algorithm

Link
Link functions used by GLM

Traits

Float

Type Definitions

Result