Struct linfa_linear::LinearRegression

pub struct LinearRegression { /* fields omitted */ }

An ordinary least squares linear regression model.

LinearRegression fits a linear model to minimize the residual sum of squares between the observed targets in the dataset, and the targets predicted by the linear approximation.

Ordinary least squares regression solves the overconstrainted model

y = Ax + b

by finding x and b which minimize the L_2 norm ||y - Ax - b||_2.

It currently uses the Moore-Penrose pseudo-inverse to solve y - b = Ax.

/// ## Examples

Here's an example on how to train a linear regression model on the diabetes dataset

use linfa::traits::{Fit, Predict};
use linfa_linear::LinearRegression;
use linfa::prelude::SingleTargetRegression;

let dataset = linfa_datasets::diabetes();
let model = LinearRegression::default().fit(&dataset).unwrap();
let pred = model.predict(&dataset);
let r2 = pred.r2(&dataset).unwrap();
println!("r2 from prediction: {}", r2);

Implementations

impl LinearRegression

Configure and fit a linear regression model

pub fn new() -> LinearRegression

Create a default linear regression model.

By default, an intercept will be fitted. To disable fitting an intercept, call .with_intercept(false) before calling .fit().

To additionally normalize the feature matrix before fitting, call fit_intercept_and_normalize() before calling fit(). The feature matrix will not be normalized by default.

pub fn with_intercept(self, with_intercept: bool) -> Self

Configure the linear regression model to fit an intercept. Defaults to true if not set.

pub fn with_intercept_and_normalize(self) -> Self

Configure the linear regression model to fit an intercept and to normalize the feature matrix before fitting it.

Normalizing the feature matrix is generally recommended to improve numeric stability unless features have already been normalized or are all within in a small range and all features are of similar size.

Normalization implies fitting an intercept.